requirements.txt
flask==2.3.3
flask-socketio==5.3.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
flask-login==0.6.3
requests==2.31.0